refactor: fix sign in out form
