+++ src/app/(frontend)/page.tsx
@@ -1,11 +1,11 @@
 import Button from "@ui/button/Button";
 import GoogleAuthButton from "@ui/button/GoogleAuthButton";
 import Footer from "@ui/footer/Footer";
+import CredentialsSignIn from "@ui/form/CredentialSignIn";
+import CredentialsSignOut from "@ui/form/CredentialSignOut";
 import StandardLayout from "@ui/layout/StandardLayout";
 
 import { auth } from "@/auth";
-
-import SignOut from "./_components/sign-in";
 
 export default async function Home() {
     const session = await auth();
@@ -23,8 +23,9 @@ export default async function Home() {
                     Profile page
                 </Button>
                 <Button variant={{ style: "google" }}>Test</Button>
-                <SignOut />
+                <CredentialsSignOut />
                 <GoogleAuthButton />
+                <CredentialsSignIn />
                 <p>{JSON.stringify(session)}</p>
             </StandardLayout>
             <Footer />
